# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SecureOps API
  version: 0.3.0
  summary: Unified DevSecOps security assessment platform
  description: |
    SecureOps turns fragmented security scanner output into one contextual
    security decision.

    **This specification is part of the API contract.** An API change without a
    corresponding change here is incomplete (CLAUDE.md §18).

    ## Current state

    Phase 5 of 14. Projects and scans can be created and read, findings are
    normalized and tracked across scans, and correlation groups them into
    issues. Risk scoring, remediation, and policy endpoints arrive in later
    phases.

    **Five scanner adapters are registered: Gitleaks, Syft, Grype, Semgrep, and
    Trivy.** A repository scan therefore means secret scanning, an SBOM,
    known-vulnerability matching, static analysis, and misconfiguration
    checking. Container images and DAST are not covered: Trivy image targets
    and OWASP ZAP are not built, which is also why the `image` and `endpoint`
    correlation keys do not appear.

    ## Authentication

    Every endpoint under `/api/v1` requires a bearer token, except
    `/api/v1/health`. See the `bearerAuth` security scheme below. This is an
    **interim** measure (`docs/adr/006-interim-bearer-token-auth.md`): it
    authenticates but does not authorize. Every valid token can reach every
    project, so this is safe only for a single-tenant deployment. Phase 11
    replaces it with real identity and RBAC.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/projects:
    post:
      tags: [projects]
      summary: Create a project
      description: |
        `environment`, `criticality`, and `internet_facing` are risk-engine
        inputs. They are set here rather than inferred, because a risk
        multiplier derived from a guess is a risk multiplier nobody can reason
        about.

        `slug` is derived from `name` when omitted. Unknown fields are
        rejected: silently ignoring a misspelled `criticality` would create a
        project at the default risk weight while the caller believed otherwise.
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: The project was created.
          headers:
            Location:
              description: The new project's URL.
              schema: { type: string }
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409":
          description: A project with that slug already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "500": { $ref: "#/components/responses/InternalError" }

    get:
      tags: [projects]
      summary: List projects
      operationId: listProjects
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}:
    get:
      tags: [projects]
      summary: Get a project
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/scans:
    get:
      tags: [scans]
      summary: List a project's scans
      description: |
        Newest first. Per-scanner results are omitted from list entries; fetch
        an individual scan for that detail.
      operationId: listProjectScans
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of scans.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanList"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/findings:
    get:
      tags: [findings]
      summary: List a project's findings
      description: |
        The canonical findings across every scan of this project, most severe
        first, then most recently seen.

        A finding is an identity that persists across scans, not a per-scan
        record: it keeps its `id` and `fingerprint` when the code it points at
        moves, and its `status` reflects whether the latest scans still report
        it. Fetch a scan's findings instead to ask what one scan saw.
      operationId: listProjectFindings
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          description: |
            Restrict to one lifecycle state. An unrecognised value is rejected
            rather than silently matching nothing.
          schema: { $ref: "#/components/schemas/FindingStatus" }
        - name: severity
          in: query
          description: Restrict to one severity.
          schema: { $ref: "#/components/schemas/FindingSeverity" }
        - name: scanner
          in: query
          description: Restrict to findings first reported by one scanner.
          schema: { type: string }
      responses:
        "200":
          description: A page of findings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingList"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/issues:
    get:
      tags: [issues]
      summary: List a project's correlated issues
      description: |
        The contextual answer, as opposed to the list of individual findings.

        An issue is several findings that share one attribute -- a
        vulnerability, a component, or a file -- treated as one problem. It is
        derived, recomputed after every scan from the project's live findings,
        and deterministic: the same findings always produce the same issues.

        Issues link findings; they never replace them. Every member carries its
        own finding id, its own severity as its scanner assessed it, and the
        evidence for its membership, so nothing is hidden behind the grouping.

        `severity` is derived from the members and may be one step above the
        worst of them when the findings span two or more domains, which
        `escalated` reports. It is a severity on the same scale as a finding's,
        **not** a risk score.
      operationId: listProjectIssues
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of correlated issues, most severe first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueList"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/risk:
    get:
      tags: [risk]
      summary: Get a project's risk score and trend
      description: |
        The single contextual number SecureOps exists to produce, computed by a
        pure deterministic engine from findings, correlated issues, and the
        project's declared context. No AI or heuristic model influences it.

        `score` runs 0 (secure) to 100 (critical) and saturates: past a point
        it stops separating projects, which is why `total` -- the aggregate
        before saturation -- is returned alongside it. A project asking "are we
        getting worse?" at the top of the scale must read `total`.

        `weights_digest` identifies the weight configuration the score was
        computed under. Weights are tunable by design, so two scores with
        different digests measure different things and a trend line drawn
        across a change is not meaningful.

        A project that has never been scored returns **404, never a score of
        zero**: "we have not assessed this" and "we assessed it and it is
        clean" are different claims, and only one is safe to act on.

        `complete` reports whether the score rests on full scanner coverage. A
        score from a `partial` scan was computed over whatever the working
        scanners found, and a failed scanner produces no findings -- which
        looks identical to an improvement. A gate must read this.

        The formula and every weight are documented in
        `docs/architecture/risk-engine.md` and ADR 019.
      operationId: getProjectRisk
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: history
          in: query
          required: false
          description: |
            How many previous scores to return in `history`, newest first. Zero
            omits the trend entirely.
          schema:
            type: integer
            minimum: 0
            maximum: 200
            default: 30
      responses:
        "200":
          description: The project's current risk score, with its trend.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Risk"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/remediation:
    get:
      tags: [remediation]
      summary: Get a project's ranked remediation plan
      description: |
        The question a risk score raises and cannot settle: not "how bad is
        this" but "what should I do first".

        The unit is an **action**, not a finding. One upgrade may resolve
        several findings reported by several scanners, and it appears once,
        with every member retained.

        Actions are ranked by `risk_removed` -- how far the project score would
        actually fall if the action were taken, computed by rerunning the risk
        engine without that action's findings. This is deliberately not the sum
        of the members' risk: aggregation is max-dominant, so removing the
        single finding setting a project's floor can outrank clearing several
        lesser ones that sum higher.

        `fixed_versions` carries every version the vendors reported and does
        **not** choose between them. Choosing needs ecosystem-specific version
        ordering -- semver, PEP 440, Maven, Debian epochs -- and a comparator
        correct for one is wrong for another.

        Every statement carries a `source`. Per §11, AI-derived content must be
        structurally distinguishable from verified data; `ai_explanation` is
        declared for that purpose and is **never produced**, because no model
        influences remediation.

        The plan is derived on read, never stored, so an action's status cannot
        drift from its members'. See
        `docs/architecture/remediation.md` and ADR 020.
      operationId: getProjectRemediation
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The project's remediation actions, most valuable first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemediationPlan"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/projects/{projectID}/policy:
    get:
      tags: [policy]
      summary: Get a project's security gate configuration
      description: |
        The rules a scan is judged against. A project that has never configured
        one is gated by the default policy rather than by nothing — an absent
        policy meaning "everything passes" would make deleting a row a way to
        silently disable the gate.
      operationId: getProjectPolicy
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The project's policy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    put:
      tags: [policy]
      summary: Replace a project's security gate configuration
      description: |
        The most security-sensitive write in this API. Raising `max` on the
        critical-findings rule from 0 to 50 turns the gate off, so every change
        is recorded in the append-only audit log **in the same transaction as
        the change itself** — a record that could be lost while the change
        survived would leave gaps that look exactly like inaction.

        The request replaces the policy wholesale; there is no partial update,
        because a half-applied gate is not a state worth being able to reach.

        **This endpoint is authenticated and not authorized.** Every valid token
        may edit every project's policy (threat model T-23, Phase 11). The audit
        log records who changed what, not whether they were entitled to —
        detection is not prevention, and the gap is stated rather than implied.
      operationId: setProjectPolicy
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Policy" }
      responses:
        "200":
          description: The policy as stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scans/{scanID}/gate:
    get:
      tags: [policy]
      summary: Get a scan's gate decision
      description: |
        `pass`, `warn`, or `fail`, with the whole of its reasoning.

        `conditions` lists **every** rule evaluated, breached or not. §12
        forbids a bare verdict, and a result carrying only breaches makes a pass
        unfalsifiable: "the project is clean" and "the policy checks nothing"
        would look identical.

        `coverage` says whether the scan had full scanner coverage. An
        incomplete scan can never produce `pass`: a scanner that crashed
        reported nothing, fewer findings breach fewer rules, and without this a
        broken scan would pass the gate precisely because it broke.

        A scan that has not been evaluated returns **404, never a pass** — "no
        gate ran" and "the gate passed" are different claims, and only one of
        them clears a change to ship.
      operationId: getScanGate
      parameters:
        - $ref: "#/components/parameters/ScanID"
      responses:
        "200":
          description: The gate decision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GateResult" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/findings/{findingID}/status:
    post:
      tags: [findings]
      summary: Change a finding's status
      description: |
        A human judgement about a finding: acknowledge it, start work, dismiss
        it as a false positive, accept it as a known risk, or reopen it.

        **`resolved` and `reopened` are refused with 422.** They mean "a
        scanner stopped reporting this" and "a scanner saw it again" —
        observations, not opinions. Allowing them by hand would make an
        asserted fix indistinguishable from a verified one, drop the risk
        score, and turn a failing gate green with nothing repaired. To record a
        fix a scanner cannot see, use `ignored` with a note saying why.

        Dismissing a finding lowers the project's risk score, removes its
        remediation work, and can change a gate verdict, so the change and its
        audit record are written in one transaction (ADR 024).

        Requires a `service` token. Deliberately not `admin`: gating triage
        behind admin would make weakening the policy the easier route to the
        same outcome.
      operationId: transitionFinding
      parameters:
        - $ref: "#/components/parameters/FindingID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransitionRequest" }
      responses:
        "200":
          description: The recorded transition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transition" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: |
            A well-formed request for a transition a person may not make —
            setting `resolved` or `reopened`. Distinct from 400 on purpose: the
            request is not malformed, so retrying it unchanged will not help.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/findings/{findingID}/history:
    get:
      tags: [findings]
      summary: Get a finding's status history
      description: |
        Every transition the finding has been through, newest first, with who
        made it, when, why, and both states — §17's requirement, served rather
        than merely stored.

        Scan-driven transitions appear alongside human ones and are attributed
        to `system`: a finding became resolved because a scan stopped reporting
        it, not because anyone decided anything.
      operationId: getFindingHistory
      parameters:
        - $ref: "#/components/parameters/FindingID"
      responses:
        "200":
          description: The finding's transitions, newest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitionHistory" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/scans/{scanID}/findings:
    get:
      tags: [findings]
      summary: List the findings one scan reported
      description: |
        Only findings this scan actually saw. A finding that exists in the
        project but was not reported by this scan does not appear, which is
        what distinguishes "what did this scan find" from "what is wrong with
        this project".
      operationId: listScanFindings
      parameters:
        - $ref: "#/components/parameters/ScanID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A page of findings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingList"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An interim static bearer token
        (`docs/adr/006-interim-bearer-token-auth.md`), configured through
        `SECUREOPS_API_TOKENS` as comma-separated `label:secret` pairs. The
        label identifies the client in the audit log.

        It authenticates; it does **not** authorize. Every valid token reaches
        every project. Tokens are compared as SHA-256 digests in constant time,
        must be at least 32 characters, and the API refuses to start without
        one. Phase 11 replaces this with real identity and RBAC.
...
  responses:
...
    PayloadTooLarge:
      description: The request body exceeded the configured cap.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...